+++ pkg/grpc/default_api/api.go
@@ -30,8 +30,6 @@ import (
 	"strings"
 	"sync"
 
-	l8_comp_pubsub "mosn.io/layotto/components/pubsub"
-
 	"github.com/dapr/components-contrib/bindings"
 
 	"mosn.io/pkg/utils"
@@ -46,10 +44,7 @@ import (
 	runtime_lock "mosn.io/layotto/pkg/runtime/lock"
 	runtime_sequencer "mosn.io/layotto/pkg/runtime/sequencer"
 
-	contrib_contenttype "github.com/dapr/components-contrib/contenttype"
 	"github.com/dapr/components-contrib/pubsub"
-	contrib_pubsub "github.com/dapr/components-contrib/pubsub"
-	"github.com/google/uuid"
 	jsoniter "github.com/json-iterator/go"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
@@ -414,74 +409,6 @@ func (a *api) SubscribeConfiguration(sub runtimev1pb.Runtime_SubscribeConfigurat
 	wg.Wait()
 	log.DefaultLogger.Warnf("subscribe gorountine exit")
 	return subErr
-}
-
-func (a *api) PublishEvent(ctx context.Context, in *runtimev1pb.PublishEventRequest) (*emptypb.Empty, error) {
-	result, err := a.doPublishEvent(ctx, in.PubsubName, in.Topic, in.Data, in.DataContentType, in.Metadata)
-	if err != nil {
-		log.DefaultLogger.Errorf("[runtime] [grpc.PublishEvent] %v", err)
-	}
-	return result, err
-}
-
-// doPublishEvent is a protocal irrelevant function to do event publishing.
-// It's easy to add APIs for other protocals.Just move this func to a separate layer if you need.
-func (a *api) doPublishEvent(ctx context.Context, pubsubName string, topic string, data []byte, contentType string, metadata map[string]string) (*emptypb.Empty, error) {
-	// 1. validate
-	if pubsubName == "" {
-		err := status.Error(codes.InvalidArgument, messages.ErrPubsubEmpty)
-		return &emptypb.Empty{}, err
-	}
-	if topic == "" {
-		err := status.Errorf(codes.InvalidArgument, messages.ErrTopicEmpty, pubsubName)
-		return &emptypb.Empty{}, err
-	}
-	// 2. get component
-	component, ok := a.pubSubs[pubsubName]
-	if !ok {
-		err := status.Errorf(codes.InvalidArgument, messages.ErrPubsubNotFound, pubsubName)
-		return &emptypb.Empty{}, err
-	}
-
-	// 3. new cloudevent request
-	if data == nil {
-		data = []byte{}
-	}
-	var envelope map[string]interface{}
-	var err error = nil
-	if contrib_contenttype.IsCloudEventContentType(contentType) {
-		envelope, err = contrib_pubsub.FromCloudEvent(data, topic, pubsubName, "")
-		if err != nil {
-			err = status.Errorf(codes.InvalidArgument, messages.ErrPubsubCloudEventCreation, err.Error())
-			return &emptypb.Empty{}, err
-		}
-	} else {
-		envelope = contrib_pubsub.NewCloudEventsEnvelope(uuid.New().String(), l8_comp_pubsub.DefaultCloudEventSource, l8_comp_pubsub.DefaultCloudEventType, "", topic, pubsubName,
-			contentType, data, "")
-	}
-	features := component.Features()
-	pubsub.ApplyMetadata(envelope, features, metadata)
-
-	b, err := jsoniter.ConfigFastest.Marshal(envelope)
-	if err != nil {
-		err = status.Errorf(codes.InvalidArgument, messages.ErrPubsubCloudEventsSer, topic, pubsubName, err.Error())
-		return &emptypb.Empty{}, err
-	}
-	// 4. publish
-	req := pubsub.PublishRequest{
-		PubsubName: pubsubName,
-		Topic:      topic,
-		Data:       b,
-		Metadata:   metadata,
-	}
-
-	// TODO limit topic scope
-	err = component.Publish(&req)
-	if err != nil {
-		nerr := status.Errorf(codes.Internal, messages.ErrPubsubPublishMessage, topic, pubsubName, err.Error())
-		return &emptypb.Empty{}, nerr
-	}
-	return &emptypb.Empty{}, nil
 }
 
 // GetState obtains the state for a specific key.
