+++ pkg/grpc/default_api/api_pubsub.go
@@ -18,18 +18,95 @@ package default_api
 
 import (
 	"context"
-	"encoding/base64"
 	"errors"
 	"fmt"
-	"github.com/dapr/components-contrib/contenttype"
-	"github.com/dapr/components-contrib/pubsub"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
+	l8_comp_pubsub "mosn.io/layotto/components/pubsub"
+
 	_ "net/http/pprof"
 
+	contrib_contenttype "github.com/dapr/components-contrib/contenttype"
+	"github.com/dapr/components-contrib/pubsub"
+	contrib_pubsub "github.com/dapr/components-contrib/pubsub"
+	"github.com/google/uuid"
+	jsoniter "github.com/json-iterator/go"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/types/known/emptypb"
+
+	"encoding/base64"
+	"github.com/dapr/components-contrib/contenttype"
+	"mosn.io/layotto/pkg/messages"
+	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
 	"mosn.io/pkg/log"
 )
+
+func (a *api) PublishEvent(ctx context.Context, in *runtimev1pb.PublishEventRequest) (*emptypb.Empty, error) {
+	result, err := a.doPublishEvent(ctx, in.PubsubName, in.Topic, in.Data, in.DataContentType, in.Metadata)
+	if err != nil {
+		log.DefaultLogger.Errorf("[runtime] [grpc.PublishEvent] %v", err)
+	}
+	return result, err
+}
+
+// doPublishEvent is a protocal irrelevant function to do event publishing.
+// It's easy to add APIs for other protocals.Just move this func to a separate layer if you need.
+func (a *api) doPublishEvent(ctx context.Context, pubsubName string, topic string, data []byte, contentType string, metadata map[string]string) (*emptypb.Empty, error) {
+	// 1. validate
+	if pubsubName == "" {
+		err := status.Error(codes.InvalidArgument, messages.ErrPubsubEmpty)
+		return &emptypb.Empty{}, err
+	}
+	if topic == "" {
+		err := status.Errorf(codes.InvalidArgument, messages.ErrTopicEmpty, pubsubName)
+		return &emptypb.Empty{}, err
+	}
+	// 2. get component
+	component, ok := a.pubSubs[pubsubName]
+	if !ok {
+		err := status.Errorf(codes.InvalidArgument, messages.ErrPubsubNotFound, pubsubName)
+		return &emptypb.Empty{}, err
+	}
+
+	// 3. new cloudevent request
+	if data == nil {
+		data = []byte{}
+	}
+	var envelope map[string]interface{}
+	var err error = nil
+	if contrib_contenttype.IsCloudEventContentType(contentType) {
+		envelope, err = contrib_pubsub.FromCloudEvent(data, topic, pubsubName, "")
+		if err != nil {
+			err = status.Errorf(codes.InvalidArgument, messages.ErrPubsubCloudEventCreation, err.Error())
+			return &emptypb.Empty{}, err
+		}
+	} else {
+		envelope = contrib_pubsub.NewCloudEventsEnvelope(uuid.New().String(), l8_comp_pubsub.DefaultCloudEventSource, l8_comp_pubsub.DefaultCloudEventType, "", topic, pubsubName,
+			contentType, data, "")
+	}
+	features := component.Features()
+	pubsub.ApplyMetadata(envelope, features, metadata)
+
+	b, err := jsoniter.ConfigFastest.Marshal(envelope)
+	if err != nil {
+		err = status.Errorf(codes.InvalidArgument, messages.ErrPubsubCloudEventsSer, topic, pubsubName, err.Error())
+		return &emptypb.Empty{}, err
+	}
+	// 4. publish
+	req := pubsub.PublishRequest{
+		PubsubName: pubsubName,
+		Topic:      topic,
+		Data:       b,
+		Metadata:   metadata,
+	}
+
+	// TODO limit topic scope
+	err = component.Publish(&req)
+	if err != nil {
+		nerr := status.Errorf(codes.Internal, messages.ErrPubsubPublishMessage, topic, pubsubName, err.Error())
+		return &emptypb.Empty{}, nerr
+	}
+	return &emptypb.Empty{}, nil
+}
 
 func (a *api) startSubscribing() error {
 	// 1. check if there is no need to do it
