+++ pkg/grpc/default_api/api_state_test.go
@@ -307,3 +307,350 @@ func TestStateItem2DeleteRequest(t *testing.T) {
 	assert.Equal(t, req.Options.Consistency, "eventual")
 	assert.Equal(t, req.Options.Concurrency, "last-write")
 }
+
+func TestGetBulkState(t *testing.T) {
+	t.Run("state store not found", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.GetBulkStateRequest{
+			StoreName: "abc",
+		}
+		_, err := api.GetBulkState(context.Background(), req)
+		assert.Equal(t, "rpc error: code = InvalidArgument desc = state store abc is not found", err.Error())
+	})
+
+	t.Run("get state error", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		mockStore.EXPECT().BulkGet(gomock.Any()).Return(false, nil, fmt.Errorf("net error"))
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.GetBulkStateRequest{
+			StoreName: "mock",
+			Keys:      []string{"mykey"},
+		}
+		_, err := api.GetBulkState(context.Background(), req)
+		assert.Equal(t, "net error", err.Error())
+	})
+
+	t.Run("support bulk get", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+
+		compResp := []state.BulkGetResponse{
+			{
+				Data:     []byte("mock data"),
+				Metadata: nil,
+			},
+		}
+		mockStore.EXPECT().BulkGet(gomock.Any()).Return(true, compResp, nil)
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.GetBulkStateRequest{
+			StoreName: "mock",
+			Keys:      []string{"mykey"},
+		}
+		rsp, err := api.GetBulkState(context.Background(), req)
+		assert.Nil(t, err)
+		assert.Equal(t, []byte("mock data"), rsp.GetItems()[0].GetData())
+	})
+
+	t.Run("don't support bulk get", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+
+		resp1 := &state.GetResponse{
+			Data:     []byte("mock data"),
+			Metadata: nil,
+		}
+
+		resp2 := &state.GetResponse{
+			Data:     []byte("mock data2"),
+			Metadata: nil,
+		}
+		mockStore.EXPECT().BulkGet(gomock.Any()).Return(false, nil, nil)
+		mockStore.EXPECT().Get(gomock.Any()).Return(resp1, nil)
+		mockStore.EXPECT().Get(gomock.Any()).Return(resp2, nil)
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.GetBulkStateRequest{
+			StoreName: "mock",
+			Keys:      []string{"mykey", "mykey2"},
+		}
+		rsp, err := api.GetBulkState(context.Background(), req)
+		assert.Nil(t, err)
+		assert.Equal(t, 2, len(rsp.GetItems()))
+	})
+
+}
+
+func TestGetState(t *testing.T) {
+	t.Run("state store not found", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.GetStateRequest{
+			StoreName: "abc",
+		}
+		_, err := api.GetState(context.Background(), req)
+		assert.Equal(t, "rpc error: code = InvalidArgument desc = state store abc is not found", err.Error())
+	})
+
+	t.Run("state store not configured", func(t *testing.T) {
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil, nil)
+		req := &runtimev1pb.GetStateRequest{
+			StoreName: "abc",
+		}
+		_, err := api.GetState(context.Background(), req)
+		assert.Equal(t, "rpc error: code = FailedPrecondition desc = state store is not configured", err.Error())
+	})
+
+	t.Run("get modified state key error", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.GetStateRequest{
+			StoreName: "mock",
+			Key:       "mykey||abc",
+		}
+		_, err := api.GetState(context.Background(), req)
+		assert.Equal(t, "input key/keyPrefix 'mykey||abc' can't contain '||'", err.Error())
+	})
+
+	t.Run("get state error", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		mockStore.EXPECT().Get(gomock.Any()).Return(nil, fmt.Errorf("net error"))
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.GetStateRequest{
+			StoreName: "mock",
+			Key:       "mykey",
+		}
+		_, err := api.GetState(context.Background(), req)
+		assert.Equal(t, "rpc error: code = Internal desc = fail to get mykey from state store mock: net error", err.Error())
+	})
+
+	t.Run("normal", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+
+		compResp := &state.GetResponse{
+			Data:     []byte("mock data"),
+			Metadata: nil,
+		}
+		mockStore.EXPECT().Get(gomock.Any()).Return(compResp, nil)
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.GetStateRequest{
+			StoreName: "mock",
+			Key:       "mykey",
+		}
+		rsp, err := api.GetState(context.Background(), req)
+		assert.Nil(t, err)
+		assert.Equal(t, []byte("mock data"), rsp.GetData())
+	})
+
+}
+
+func TestDeleteState(t *testing.T) {
+	t.Run("normal", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		mockStore.EXPECT().Delete(gomock.Any()).DoAndReturn(func(req *state.DeleteRequest) error {
+			assert.Equal(t, "abc", req.Key)
+			return nil
+		})
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.DeleteStateRequest{
+			StoreName: "mock",
+			Key:       "abc",
+		}
+		_, err := api.DeleteState(context.Background(), req)
+		assert.Nil(t, err)
+	})
+
+	t.Run("net error", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		mockStore.EXPECT().Delete(gomock.Any()).Return(fmt.Errorf("net error"))
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.DeleteStateRequest{
+			StoreName: "mock",
+			Key:       "abc",
+		}
+		_, err := api.DeleteState(context.Background(), req)
+		assert.NotNil(t, err)
+		assert.Equal(t, "rpc error: code = Internal desc = failed deleting state with key abc: net error", err.Error())
+	})
+}
+
+func TestDeleteBulkState(t *testing.T) {
+	t.Run("normal", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		mockStore.EXPECT().BulkDelete(gomock.Any()).DoAndReturn(func(reqs []state.DeleteRequest) error {
+			assert.Equal(t, "abc", reqs[0].Key)
+			return nil
+		})
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.DeleteBulkStateRequest{
+			StoreName: "mock",
+			States: []*runtimev1pb.StateItem{
+				{
+					Key: "abc",
+				},
+			},
+		}
+		_, err := api.DeleteBulkState(context.Background(), req)
+		assert.Nil(t, err)
+	})
+
+	t.Run("net error", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		mockStore.EXPECT().BulkDelete(gomock.Any()).Return(fmt.Errorf("net error"))
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.DeleteBulkStateRequest{
+			StoreName: "mock",
+			States: []*runtimev1pb.StateItem{
+				{
+					Key: "abc",
+				},
+			},
+		}
+		_, err := api.DeleteBulkState(context.Background(), req)
+		assert.NotNil(t, err)
+		assert.Equal(t, "net error", err.Error())
+	})
+}
+
+type MockTxStore struct {
+	state.Store
+	state.TransactionalStore
+}
+
+func (m *MockTxStore) Init(metadata state.Metadata) error {
+	return m.Store.Init(metadata)
+}
+
+func TestExecuteStateTransaction(t *testing.T) {
+	t.Run("state store not found", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.ExecuteStateTransactionRequest{
+			StoreName: "abc",
+		}
+		_, err := api.ExecuteStateTransaction(context.Background(), req)
+		assert.Equal(t, "rpc error: code = InvalidArgument desc = state store abc is not found", err.Error())
+	})
+
+	t.Run("state store not configured", func(t *testing.T) {
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil, nil)
+		req := &runtimev1pb.ExecuteStateTransactionRequest{
+			StoreName: "abc",
+		}
+		_, err := api.ExecuteStateTransaction(context.Background(), req)
+		assert.Equal(t, "rpc error: code = FailedPrecondition desc = state store is not configured", err.Error())
+	})
+
+	t.Run("normal", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return([]state.Feature{state.FeatureTransactional})
+
+		mockTxStore := mock_state.NewMockTransactionalStore(gomock.NewController(t))
+		mockTxStore.EXPECT().Multi(gomock.Any()).DoAndReturn(func(req *state.TransactionalStateRequest) error {
+			assert.Equal(t, 2, len(req.Operations))
+			assert.Equal(t, "mosn", req.Metadata["runtime"])
+			assert.Equal(t, state.Upsert, req.Operations[0].Operation)
+			assert.Equal(t, state.Delete, req.Operations[1].Operation)
+			return nil
+		})
+
+		store := &MockTxStore{
+			mockStore,
+			mockTxStore,
+		}
+
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": store}, nil, nil, nil, nil)
+		req := &runtimev1pb.ExecuteStateTransactionRequest{
+			StoreName: "mock",
+			Operations: []*runtimev1pb.TransactionalStateOperation{
+				{
+					OperationType: string(state.Upsert),
+					Request: &runtimev1pb.StateItem{
+						Key:   "upsert",
+						Value: []byte("mock data"),
+					},
+				},
+				{
+					OperationType: string(state.Delete),
+					Request: &runtimev1pb.StateItem{
+						Key: "delete_abc",
+					},
+				},
+				{
+					OperationType: string(state.Delete),
+				},
+			},
+			Metadata: map[string]string{
+				"runtime": "mosn",
+			},
+		}
+		_, err := api.ExecuteStateTransaction(context.Background(), req)
+		assert.Nil(t, err)
+	})
+
+	t.Run("net error", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return([]state.Feature{state.FeatureTransactional})
+
+		mockTxStore := mock_state.NewMockTransactionalStore(gomock.NewController(t))
+		mockTxStore.EXPECT().Multi(gomock.Any()).Return(fmt.Errorf("net error"))
+
+		store := &MockTxStore{
+			mockStore,
+			mockTxStore,
+		}
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": store}, nil, nil, nil, nil)
+		req := &runtimev1pb.ExecuteStateTransactionRequest{
+			StoreName: "mock",
+			Operations: []*runtimev1pb.TransactionalStateOperation{
+				{
+					OperationType: string(state.Upsert),
+					Request: &runtimev1pb.StateItem{
+						Key:   "upsert",
+						Value: []byte("mock data"),
+					},
+				},
+				{
+					OperationType: string(state.Delete),
+					Request: &runtimev1pb.StateItem{
+						Key: "delete_abc",
+					},
+				},
+				{
+					OperationType: string(state.Delete),
+				},
+			},
+			Metadata: map[string]string{
+				"runtime": "mosn",
+			},
+		}
+		_, err := api.ExecuteStateTransaction(context.Background(), req)
+		assert.NotNil(t, err)
+		assert.Equal(t, "rpc error: code = Internal desc = error while executing state transaction: net error", err.Error())
+	})
+}
