+++ pkg/grpc/default_api/api_test.go
@@ -18,23 +18,17 @@ package default_api
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"github.com/dapr/components-contrib/bindings"
-	"github.com/dapr/components-contrib/pubsub"
-	"github.com/dapr/components-contrib/state"
 	"github.com/golang/mock/gomock"
 	"github.com/phayes/freeport"
 	"github.com/stretchr/testify/assert"
-	rawGRPC "google.golang.org/grpc"
-	"google.golang.org/grpc/test/bufconn"
 	l8grpc "mosn.io/layotto/pkg/grpc"
 	"mosn.io/layotto/pkg/mock/runtime"
 	"net"
 	"testing"
 
 	"errors"
-	jsoniter "github.com/json-iterator/go"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"io"
@@ -46,10 +40,7 @@ import (
 	"mosn.io/layotto/pkg/mock"
 	mock_invoker "mosn.io/layotto/pkg/mock/components/invoker"
 	mock_lock "mosn.io/layotto/pkg/mock/components/lock"
-	mock_pubsub "mosn.io/layotto/pkg/mock/components/pubsub"
 	mock_sequencer "mosn.io/layotto/pkg/mock/components/sequencer"
-	mock_state "mosn.io/layotto/pkg/mock/components/state"
-	mock_appcallback "mosn.io/layotto/pkg/mock/runtime/appcallback"
 	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
 
 	"time"
@@ -139,58 +130,6 @@ func TestSayHello(t *testing.T) {
 		if err != ErrNoInstance {
 			t.Fatalf("expected got a no instance error, but got %v", err)
 		}
-	})
-}
-
-func TestMosnRuntime_publishMessageGRPC(t *testing.T) {
-	t.Run("publish success", func(t *testing.T) {
-		subResp := &runtimev1pb.TopicEventResponse{
-			Status: runtimev1pb.TopicEventResponse_SUCCESS,
-		}
-		// init grpc server
-		mockAppCallbackServer := mock_appcallback.NewMockAppCallbackServer(gomock.NewController(t))
-		mockAppCallbackServer.EXPECT().OnTopicEvent(gomock.Any(), gomock.Any()).Return(subResp, nil)
-
-		lis := bufconn.Listen(1024 * 1024)
-		s := grpc.NewServer()
-		runtimev1pb.RegisterAppCallbackServer(s, mockAppCallbackServer)
-		go func() {
-			s.Serve(lis)
-		}()
-
-		// init callback client
-		callbackClient, err := grpc.DialContext(context.Background(), "bufnet", rawGRPC.WithInsecure(), rawGRPC.WithContextDialer(func(ctx context.Context, s string) (net.Conn, error) {
-			return lis.Dial()
-		}))
-		assert.Nil(t, err)
-
-		cloudEvent := map[string]interface{}{
-			pubsub.IDField:              "id",
-			pubsub.SourceField:          "source",
-			pubsub.DataContentTypeField: "content-type",
-			pubsub.TypeField:            "type",
-			pubsub.SpecVersionField:     "v1.0.0",
-			pubsub.DataBase64Field:      "bGF5b3R0bw==",
-		}
-
-		data, err := json.Marshal(cloudEvent)
-		assert.Nil(t, err)
-
-		msg := &pubsub.NewMessage{
-			Data:     data,
-			Topic:    "layotto",
-			Metadata: make(map[string]string),
-		}
-		a := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil, nil)
-
-		var apiForTest = a.(*api)
-		//apiForTest.errInt = func(err error, format string, args ...interface{}) {
-		//	log.DefaultLogger.Errorf("[runtime] occurs an error: "+err.Error()+", "+format, args...)
-		//}
-		apiForTest.AppCallbackConn = callbackClient
-		apiForTest.json = jsoniter.ConfigFastest
-		err = apiForTest.publishMessageGRPC(context.Background(), msg)
-		assert.Nil(t, err)
 	})
 }
 
@@ -371,415 +310,6 @@ func TestInvokeService(t *testing.T) {
 
 		_, err := a.InvokeService(context.Background(), in)
 		assert.Nil(t, err)
-	})
-}
-
-func TestPublishEvent(t *testing.T) {
-	t.Run("invalid pubsub name", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockPubSub := mock_pubsub.NewMockPubSub(ctrl)
-		api := NewAPI("", nil, nil, nil, map[string]pubsub.PubSub{"mock": mockPubSub}, nil, nil, nil, nil, nil)
-		_, err := api.PublishEvent(context.Background(), &runtimev1pb.PublishEventRequest{})
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = pubsub name is empty", err.Error())
-	})
-
-	t.Run("invalid topic", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockPubSub := mock_pubsub.NewMockPubSub(ctrl)
-		api := NewAPI("", nil, nil, nil, map[string]pubsub.PubSub{"mock": mockPubSub}, nil, nil, nil, nil, nil)
-		req := &runtimev1pb.PublishEventRequest{
-			PubsubName: "abc",
-		}
-		_, err := api.PublishEvent(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = topic is empty in pubsub abc", err.Error())
-	})
-
-	t.Run("component not found", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockPubSub := mock_pubsub.NewMockPubSub(ctrl)
-		api := NewAPI("", nil, nil, nil, map[string]pubsub.PubSub{"mock": mockPubSub}, nil, nil, nil, nil, nil)
-		req := &runtimev1pb.PublishEventRequest{
-			PubsubName: "abc",
-			Topic:      "abc",
-		}
-		_, err := api.PublishEvent(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = pubsub abc not found", err.Error())
-	})
-
-	t.Run("publish success", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockPubSub := mock_pubsub.NewMockPubSub(ctrl)
-		mockPubSub.EXPECT().Publish(gomock.Any()).Return(nil)
-		mockPubSub.EXPECT().Features().Return(nil)
-		api := NewAPI("", nil, nil, nil, map[string]pubsub.PubSub{"mock": mockPubSub}, nil, nil, nil, nil, nil)
-		req := &runtimev1pb.PublishEventRequest{
-			PubsubName: "mock",
-			Topic:      "abc",
-		}
-		_, err := api.PublishEvent(context.Background(), req)
-		assert.Nil(t, err)
-	})
-
-	t.Run("publish net error", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockPubSub := mock_pubsub.NewMockPubSub(ctrl)
-		mockPubSub.EXPECT().Publish(gomock.Any()).Return(fmt.Errorf("net error"))
-		mockPubSub.EXPECT().Features().Return(nil)
-		api := NewAPI("", nil, nil, nil, map[string]pubsub.PubSub{"mock": mockPubSub}, nil, nil, nil, nil, nil)
-		req := &runtimev1pb.PublishEventRequest{
-			PubsubName: "mock",
-			Topic:      "abc",
-		}
-		_, err := api.PublishEvent(context.Background(), req)
-		assert.NotNil(t, err)
-		assert.Equal(t, "rpc error: code = Internal desc = error when publish to topic abc in pubsub mock: net error", err.Error())
-	})
-}
-
-func TestGetBulkState(t *testing.T) {
-	t.Run("state store not found", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.GetBulkStateRequest{
-			StoreName: "abc",
-		}
-		_, err := api.GetBulkState(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = state store abc is not found", err.Error())
-	})
-
-	t.Run("get state error", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-		mockStore.EXPECT().BulkGet(gomock.Any()).Return(false, nil, fmt.Errorf("net error"))
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.GetBulkStateRequest{
-			StoreName: "mock",
-			Keys:      []string{"mykey"},
-		}
-		_, err := api.GetBulkState(context.Background(), req)
-		assert.Equal(t, "net error", err.Error())
-	})
-
-	t.Run("support bulk get", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-
-		compResp := []state.BulkGetResponse{
-			{
-				Data:     []byte("mock data"),
-				Metadata: nil,
-			},
-		}
-		mockStore.EXPECT().BulkGet(gomock.Any()).Return(true, compResp, nil)
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.GetBulkStateRequest{
-			StoreName: "mock",
-			Keys:      []string{"mykey"},
-		}
-		rsp, err := api.GetBulkState(context.Background(), req)
-		assert.Nil(t, err)
-		assert.Equal(t, []byte("mock data"), rsp.GetItems()[0].GetData())
-	})
-
-	t.Run("don't support bulk get", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-
-		resp1 := &state.GetResponse{
-			Data:     []byte("mock data"),
-			Metadata: nil,
-		}
-
-		resp2 := &state.GetResponse{
-			Data:     []byte("mock data2"),
-			Metadata: nil,
-		}
-		mockStore.EXPECT().BulkGet(gomock.Any()).Return(false, nil, nil)
-		mockStore.EXPECT().Get(gomock.Any()).Return(resp1, nil)
-		mockStore.EXPECT().Get(gomock.Any()).Return(resp2, nil)
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.GetBulkStateRequest{
-			StoreName: "mock",
-			Keys:      []string{"mykey", "mykey2"},
-		}
-		rsp, err := api.GetBulkState(context.Background(), req)
-		assert.Nil(t, err)
-		assert.Equal(t, 2, len(rsp.GetItems()))
-	})
-
-}
-
-func TestGetState(t *testing.T) {
-	t.Run("state store not found", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.GetStateRequest{
-			StoreName: "abc",
-		}
-		_, err := api.GetState(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = state store abc is not found", err.Error())
-	})
-
-	t.Run("state store not configured", func(t *testing.T) {
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil, nil)
-		req := &runtimev1pb.GetStateRequest{
-			StoreName: "abc",
-		}
-		_, err := api.GetState(context.Background(), req)
-		assert.Equal(t, "rpc error: code = FailedPrecondition desc = state store is not configured", err.Error())
-	})
-
-	t.Run("get modified state key error", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.GetStateRequest{
-			StoreName: "mock",
-			Key:       "mykey||abc",
-		}
-		_, err := api.GetState(context.Background(), req)
-		assert.Equal(t, "input key/keyPrefix 'mykey||abc' can't contain '||'", err.Error())
-	})
-
-	t.Run("get state error", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-		mockStore.EXPECT().Get(gomock.Any()).Return(nil, fmt.Errorf("net error"))
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.GetStateRequest{
-			StoreName: "mock",
-			Key:       "mykey",
-		}
-		_, err := api.GetState(context.Background(), req)
-		assert.Equal(t, "rpc error: code = Internal desc = fail to get mykey from state store mock: net error", err.Error())
-	})
-
-	t.Run("normal", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-
-		compResp := &state.GetResponse{
-			Data:     []byte("mock data"),
-			Metadata: nil,
-		}
-		mockStore.EXPECT().Get(gomock.Any()).Return(compResp, nil)
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.GetStateRequest{
-			StoreName: "mock",
-			Key:       "mykey",
-		}
-		rsp, err := api.GetState(context.Background(), req)
-		assert.Nil(t, err)
-		assert.Equal(t, []byte("mock data"), rsp.GetData())
-	})
-
-}
-
-func TestDeleteState(t *testing.T) {
-	t.Run("normal", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-		mockStore.EXPECT().Delete(gomock.Any()).DoAndReturn(func(req *state.DeleteRequest) error {
-			assert.Equal(t, "abc", req.Key)
-			return nil
-		})
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.DeleteStateRequest{
-			StoreName: "mock",
-			Key:       "abc",
-		}
-		_, err := api.DeleteState(context.Background(), req)
-		assert.Nil(t, err)
-	})
-
-	t.Run("net error", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-		mockStore.EXPECT().Delete(gomock.Any()).Return(fmt.Errorf("net error"))
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.DeleteStateRequest{
-			StoreName: "mock",
-			Key:       "abc",
-		}
-		_, err := api.DeleteState(context.Background(), req)
-		assert.NotNil(t, err)
-		assert.Equal(t, "rpc error: code = Internal desc = failed deleting state with key abc: net error", err.Error())
-	})
-}
-
-func TestDeleteBulkState(t *testing.T) {
-	t.Run("normal", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-		mockStore.EXPECT().BulkDelete(gomock.Any()).DoAndReturn(func(reqs []state.DeleteRequest) error {
-			assert.Equal(t, "abc", reqs[0].Key)
-			return nil
-		})
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.DeleteBulkStateRequest{
-			StoreName: "mock",
-			States: []*runtimev1pb.StateItem{
-				{
-					Key: "abc",
-				},
-			},
-		}
-		_, err := api.DeleteBulkState(context.Background(), req)
-		assert.Nil(t, err)
-	})
-
-	t.Run("net error", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-		mockStore.EXPECT().BulkDelete(gomock.Any()).Return(fmt.Errorf("net error"))
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.DeleteBulkStateRequest{
-			StoreName: "mock",
-			States: []*runtimev1pb.StateItem{
-				{
-					Key: "abc",
-				},
-			},
-		}
-		_, err := api.DeleteBulkState(context.Background(), req)
-		assert.NotNil(t, err)
-		assert.Equal(t, "net error", err.Error())
-	})
-}
-
-type MockTxStore struct {
-	state.Store
-	state.TransactionalStore
-}
-
-func (m *MockTxStore) Init(metadata state.Metadata) error {
-	return m.Store.Init(metadata)
-}
-
-func TestExecuteStateTransaction(t *testing.T) {
-	t.Run("state store not found", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return(nil)
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
-		req := &runtimev1pb.ExecuteStateTransactionRequest{
-			StoreName: "abc",
-		}
-		_, err := api.ExecuteStateTransaction(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = state store abc is not found", err.Error())
-	})
-
-	t.Run("state store not configured", func(t *testing.T) {
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil, nil)
-		req := &runtimev1pb.ExecuteStateTransactionRequest{
-			StoreName: "abc",
-		}
-		_, err := api.ExecuteStateTransaction(context.Background(), req)
-		assert.Equal(t, "rpc error: code = FailedPrecondition desc = state store is not configured", err.Error())
-	})
-
-	t.Run("normal", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return([]state.Feature{state.FeatureTransactional})
-
-		mockTxStore := mock_state.NewMockTransactionalStore(gomock.NewController(t))
-		mockTxStore.EXPECT().Multi(gomock.Any()).DoAndReturn(func(req *state.TransactionalStateRequest) error {
-			assert.Equal(t, 2, len(req.Operations))
-			assert.Equal(t, "mosn", req.Metadata["runtime"])
-			assert.Equal(t, state.Upsert, req.Operations[0].Operation)
-			assert.Equal(t, state.Delete, req.Operations[1].Operation)
-			return nil
-		})
-
-		store := &MockTxStore{
-			mockStore,
-			mockTxStore,
-		}
-
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": store}, nil, nil, nil, nil)
-		req := &runtimev1pb.ExecuteStateTransactionRequest{
-			StoreName: "mock",
-			Operations: []*runtimev1pb.TransactionalStateOperation{
-				{
-					OperationType: string(state.Upsert),
-					Request: &runtimev1pb.StateItem{
-						Key:   "upsert",
-						Value: []byte("mock data"),
-					},
-				},
-				{
-					OperationType: string(state.Delete),
-					Request: &runtimev1pb.StateItem{
-						Key: "delete_abc",
-					},
-				},
-				{
-					OperationType: string(state.Delete),
-				},
-			},
-			Metadata: map[string]string{
-				"runtime": "mosn",
-			},
-		}
-		_, err := api.ExecuteStateTransaction(context.Background(), req)
-		assert.Nil(t, err)
-	})
-
-	t.Run("net error", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockStore := mock_state.NewMockStore(ctrl)
-		mockStore.EXPECT().Features().Return([]state.Feature{state.FeatureTransactional})
-
-		mockTxStore := mock_state.NewMockTransactionalStore(gomock.NewController(t))
-		mockTxStore.EXPECT().Multi(gomock.Any()).Return(fmt.Errorf("net error"))
-
-		store := &MockTxStore{
-			mockStore,
-			mockTxStore,
-		}
-		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": store}, nil, nil, nil, nil)
-		req := &runtimev1pb.ExecuteStateTransactionRequest{
-			StoreName: "mock",
-			Operations: []*runtimev1pb.TransactionalStateOperation{
-				{
-					OperationType: string(state.Upsert),
-					Request: &runtimev1pb.StateItem{
-						Key:   "upsert",
-						Value: []byte("mock data"),
-					},
-				},
-				{
-					OperationType: string(state.Delete),
-					Request: &runtimev1pb.StateItem{
-						Key: "delete_abc",
-					},
-				},
-				{
-					OperationType: string(state.Delete),
-				},
-			},
-			Metadata: map[string]string{
-				"runtime": "mosn",
-			},
-		}
-		_, err := api.ExecuteStateTransaction(context.Background(), req)
-		assert.NotNil(t, err)
-		assert.Equal(t, "rpc error: code = Internal desc = error while executing state transaction: net error", err.Error())
 	})
 }
